+++ src/engine/search_mtdf.h
@@ -4,7 +4,7 @@
 #include "state.h"
 #include "search.h"
 
-int SEARCH_mtdf(const chess_state_t *s, search_state_t *search_state, short depth, move_t *move, int guess);
-int SEARCH_mtdf_iterative(const chess_state_t *s, search_state_t *search_state, move_t *move);
+short SEARCH_mtdf(const chess_state_t *s, search_state_t *search_state, short depth, move_t *move, short guess);
+short SEARCH_mtdf_iterative(const chess_state_t *s, search_state_t *search_state, move_t *move);
 
 #endif
